Added note for the root option
The root option only works if the path of a request does not start with a forward slash. This was not documented and is not obvious.

It would probably also make sense to document each option that is available for configuration.

I also just filed [this bug](https://github.com/pagekit/vue-resource/issues/487) that setting configuration options inside the Vue component options doesn't work. Don't know if it should be in the documentation or not.